requirements.txt
python-dotenv==1.0.0
fastapi==0.115.8
uvicorn==0.34.0
pymongo==4.11.1
feedparser==6.0.11
schedule==1.2.2
APScheduler==3.10.4
requests==2.31.0
python-dateutil==2.8.2
langdetect==1.0.9
